[chore] #105 Organize folder in design system
