+++ feature/mypage/src/main/java/com/cmc/curtaincall/feature/mypage/MyPageScreen.kt
@@ -32,7 +32,7 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
-import com.cmc.curtaincall.common.design.component.TopAppBarOnlyAction
+import com.cmc.curtaincall.common.design.component.basic.TopAppBarOnlyAction
 import com.cmc.curtaincall.common.design.theme.Nero
 import com.cmc.curtaincall.common.design.theme.White
 import com.cmc.curtaincall.common.design.R
