+++ feature/partymember/src/main/java/com/cmc/curtaincall/feature/partymember/ui/detail/PartyMemberDetailScreen.kt
@@ -21,7 +21,7 @@ import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.cmc.curtaincall.common.design.R
-import com.cmc.curtaincall.common.design.component.TopAppBarWithAction
+import com.cmc.curtaincall.common.design.component.basic.TopAppBarWithAction
 import com.cmc.curtaincall.common.design.theme.*
 import com.cmc.curtaincall.feature.partymember.PartyType
 import com.google.accompanist.systemuicontroller.rememberSystemUiController
